pyproject.toml
[project]
name = "SplitTasks"
version = "0.0.1-alpha"
description = "SplitTasks - Manage Tasks, Expenses and Responsibilites. Available on Web, Android and iOS."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.120.2",
    "phonenumbers>=9.0.17",
    "psycopg2>=2.9.11",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.6",
    "pydantic-settings>=2.11.0",
    "pyjwt[crypto]>=2.10.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "typing-extensions>=4.15.0",
]
authors = [{ name = 'Aditya Mayukh Som', email = "aditya.m.som@axantial.com" }]
maintainers = [
    { name = 'Aditya Mayukh Som', email = "aditya.m.som@axantial.com" },
]

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-cov>=7.0.0"]

[tool.pytest.ini_options]
addopts = "-s --cov=app"

[tool.ruff]
line-length = 120

[tool.ruff.lint.isort]
length-sort-straight = true
force-single-line = false
combine-as-imports = true
